+++ HISTORY.md
@@ -1,5 +1,7 @@
 # History
 
+-   Add `mkdirp`, `readEntireDirectory`, and `deleteEntireDirectory` with fallbacks for older Node.js versions
+
 ## v1.5.0 2023 November 13
 
 -   Updated dependencies, [base files](https://github.com/bevry/base), and [editions](https://editions.bevry.me) using [boundation](https://github.com/bevry/boundation)